+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Ability to disable an `Action` after its last defined frame. See `? Action` and the keyword `; keep`
 - Moved from `Translation`, `Scaling` and `Rotation` to `anim_translate` etc
 - Changed `Rel` to `RFrames` and added `GFrames` for defining actions with global frames
+- A warning is shown if some frames don't have a background
 
 ## v0.2.2 (20th of October 2020)
 - Ability to change a keyword using `change`